pyproject.toml
[tool.poetry]
name = "civic-assistant-team-5"
version = "0.1.0"
description = "Multi Agent Real Estate Assistant (MAREA)"
authors = ["Carlos Huaccho <huaccho.carlos0612@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
fastmcp = ">=2.12.3,<3.0.0"
langchain-openai = ">=0.3.33,<0.4.0"
python-dotenv = ">=1.1.1,<2.0.0"
pytest = ">=8.4.2,<9.0.0"
ruff = ">=0.13.0,<0.14.0"
langgraph = ">=0.6.7,<0.7.0"
langchain = ">=0.3.27,<0.4.0"
langchain-core = "*"
httpx = ">=0.28.1,<0.29.0"
langchain-mcp-adapters = ">=0.1.9,<0.2.0"
rich = ">=14.1.0,<15.0.0"
mcp = "*"
gradio = ">=5.46.0,<6.0.0"
gradio_client = ">=1.13.1,<2.0.0"
fastapi = ">=0.100.0,<1.0.0"
uvicorn = ">=0.20.0,<1.0.0"
openai = ">=1.0.0,<2.0.0"
remote-pdb = "^2.1.0"
langchain-google-genai = "^2.1.12"

[tool.pytest.ini_options]
pythonpath = ["."]
